make it possible to create test-local backing services
Test local backing services are cleaned up at the end of tests.

Also changes the endpoint methods so that they use the docker network IP instead
of a port bound to localhost. This change was made so that the endpoints become
usable for communication between backing containers.

The use-case is having Postgres, Minio and Elephant repository as backing
services for elephant-index integration tests.